+++ uptrace/config.go
@@ -2,6 +2,7 @@ package uptrace
 
 import (
 	"context"
+	"os"
 
 	"github.com/uptrace/uptrace-go/spanexp"
 
@@ -12,51 +13,45 @@ import (
 	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 )
 
-type Config struct {
-	// DSN is a data source name that is used to connect to uptrace.dev.
-	// Example: https://<token>@api.uptrace.dev/<project_id>
-	// The default is to use UPTRACE_DSN environment var.
+type config struct {
 	DSN string
 
-	// `service.name` resource attribute. It is merged with Config.Resource.
-	// For example, `myservice`.
-	ServiceName string
-	// `service.version` resource attribute. It is merged with Config.Resource.
-	// For example, `1.0.0`.
-	ServiceVersion string
-	// Any other resource attributes. They are merged with Config.Resource.
-	//
-	// You can also use `OTEL_RESOURCE_ATTRIBUTES` env var. For example,
-	// `service.name=myservice,service.version=1.0.0`.
+	// Common options
+
+	ServiceName        string
+	ServiceVersion     string
 	ResourceAttributes []attribute.KeyValue
-	// Resource contains attributes representing an entity that produces telemetry.
-	// Resource attributes are copied to all spans and events.
-	//
-	// The default is `resource.New`.
-	Resource *resource.Resource
+	Resource           *resource.Resource
 
-	// Global TextMapPropagator used by OpenTelemetry.
-	// The default is propagation.TraceContext and propagation.Baggage.
+	// Tracing options
+
+	TracingDisabled   bool
 	TextMapPropagator propagation.TextMapPropagator
+	TracerProvider    *sdktrace.TracerProvider
+	TraceSampler      sdktrace.Sampler
+	BeforeSendSpan    func(*spanexp.Span)
+	PrettyPrint       bool
 
-	// Sampler is the default sampler used when creating new spans.
-	Sampler sdktrace.Sampler
+	// Metrics options
 
-	// MetricsDisabled can be used to skip metrics configuration.
 	MetricsDisabled bool
-
-	// A hook that is called before sending a span.
-	BeforeSpanSend func(*spanexp.Span)
-
-	// PrettyPrint pretty prints spans to the stdout.
-	PrettyPrint bool
-
-	// TracerProvider overwrites the default Uptrace tracer provider.
-	// It can be used to configure Uptrace client to use OTLP exporter.
-	TracerProvider *sdktrace.TracerProvider
 }
 
-func (cfg *Config) newResource() *resource.Resource {
+func newConfig(opts []Option) *config {
+	cfg := new(config)
+
+	if dsn, ok := os.LookupEnv("UPTRACE_DSN"); ok {
+		cfg.DSN = dsn
+	}
+
+	for _, opt := range opts {
+		opt.apply(cfg)
+	}
+
+	return cfg
+}
+
+func (cfg *config) newResource() *resource.Resource {
 	return buildResource(
 		cfg.Resource, cfg.ResourceAttributes, cfg.ServiceName, cfg.ServiceVersion)
 }
@@ -98,4 +93,149 @@ func buildResource(
 	}
 
 	return res
+}
+
+//------------------------------------------------------------------------------
+
+type Option interface {
+	apply(cfg *config)
+}
+
+type option func(cfg *config)
+
+func (fn option) apply(cfg *config) {
+	fn(cfg)
+}
+
+// WithDSN configures a data source name that is used to connect to Uptrace, for example,
+// `https://<token>@api.uptrace.dev/<project_id>`.
+//
+// The default is to use UPTRACE_DSN environment variable.
+func WithDSN(dsn string) Option {
+	return option(func(cfg *config) {
+		cfg.DSN = dsn
+	})
+}
+
+// WithServiceVersion configures a `service.name` resource attribute.
+// You can use this option together with other options that configure resource attributes.
+func WithServiceName(serviceName string) Option {
+	return option(func(cfg *config) {
+		cfg.ServiceName = serviceName
+	})
+}
+
+// WithServiceVersion configures a `service.version` resource attribute, for example, `1.0.0`.
+// You can use this option together with other options that configure resource attributes.
+func WithServiceVersion(serviceVersion string) Option {
+	return option(func(cfg *config) {
+		cfg.ServiceVersion = serviceVersion
+	})
+}
+
+// WithResourceAttributes configures resource attributes that describe an entity that produces
+// telemetry, for example, such attributes as host.name, service.name, etc.
+// You can use this option together with other options that configure resource attributes.
+//
+// The default is to use `OTEL_RESOURCE_ATTRIBUTES` env var, for example,
+// `OTEL_RESOURCE_ATTRIBUTES=service.name=myservice,service.version=1.0.0`.
+func WithResourceAttributes(resourceAttributes []attribute.KeyValue) Option {
+	return option(func(cfg *config) {
+		cfg.ResourceAttributes = resourceAttributes
+	})
+}
+
+// WithResource configures a resource that describes an entity that produces telemetry,
+// for example, such attributes as host.name and service.name. All produced spans and metrics
+// with have these attributes
+func WithResource(resource *resource.Resource) Option {
+	return option(func(cfg *config) {
+		cfg.Resource = resource
+	})
+}
+
+//------------------------------------------------------------------------------
+
+type TracingOption interface {
+	Option
+	tracing()
+}
+
+type tracingOption func(cfg *config)
+
+var _ TracingOption = (*tracingOption)(nil)
+
+func (fn tracingOption) apply(cfg *config) {
+	fn(cfg)
+}
+
+func (fn tracingOption) tracing() {}
+
+// TracingDisabled can be used to skip tracing configuration.
+func WithTracingDisabled() TracingOption {
+	return tracingOption(func(cfg *config) {
+		cfg.TracingDisabled = true
+	})
+}
+
+// TracerProvider overwrites the default Uptrace tracer provider.
+// You can use it to configure Uptrace distro to use OTLP exporter.
+func WithTracerProvider(provider *sdktrace.TracerProvider) TracingOption {
+	return tracingOption(func(cfg *config) {
+		cfg.TracerProvider = provider
+	})
+}
+
+// WithTraceSampler configures a span sampler.
+func WithTraceSampler(sampler sdktrace.Sampler) TracingOption {
+	return tracingOption(func(cfg *config) {
+		cfg.TraceSampler = sampler
+	})
+}
+
+// WithTextMapPropagator sets the global TextMapPropagator used by OpenTelemetry.
+// The default is propagation.TraceContext and propagation.Baggage.
+func WithTextMapPropagator(propagator propagation.TextMapPropagator) TracingOption {
+	return tracingOption(func(cfg *config) {
+		cfg.TextMapPropagator = propagator
+	})
+}
+
+// WithBeforeSendSpan sets a hook that is called before sending a span.
+func WithBeforeSendSpan(fn func(span *spanexp.Span)) TracingOption {
+	return tracingOption(func(cfg *config) {
+		cfg.BeforeSendSpan = fn
+	})
+}
+
+// WithPrettyPrintSpanExporter adds a span exproter that prints spans to stdout.
+// It is useful for debugging or demonstration purposes.
+func WithPrettyPrintSpanExporter() TracingOption {
+	return tracingOption(func(cfg *config) {
+		cfg.PrettyPrint = true
+	})
+}
+
+//------------------------------------------------------------------------------
+
+type MetricsOption interface {
+	Option
+	metrics()
+}
+
+type metricsOption func(cfg *config)
+
+var _ MetricsOption = (*metricsOption)(nil)
+
+func (fn metricsOption) apply(cfg *config) {
+	fn(cfg)
+}
+
+func (fn metricsOption) metrics() {}
+
+// WithMetricsDisabled can be used to skip metrics configuration.
+func WithMetricsDisabled() MetricsOption {
+	return metricsOption(func(cfg *config) {
+		cfg.MetricsDisabled = true
+	})
 }